add product categories block component
The block is pass a section alias or by default is directory.  This queries the given section and generates a list of child section ids.  The child section ids are then used to filter the contents scheudls and generate a list of section names or in this case Product Categories.